fix warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@263563 c90b9560-bf6c-de11-be94-00142212c4b1

+++ sapi/cli/php_cli.c
@@ -296,7 +296,7 @@ static char *script_filename = "";
 
 static void sapi_cli_register_variables(zval *track_vars_array TSRMLS_DC) /* {{{ */
 {
-	size_t len;
+	unsigned int len;
 	char   *docroot = "";
 
 	/* In CGI mode, we consider the environment to be a part of the server
